pyproject.toml
[tool.poetry]
name = "emails-service"
version = "0.1.0"
description = "Small microservice for sending emails."
authors = ["Andrii Isiuk"]

[tool.poetry.dependencies]
python = "^3.7"
djangorestframework = "^3.11.0"
environs = "^7.3.0"
psycopg2-binary = "^2.8.5"
django-cleanup = "^4.0.0"
celery = "^4.4.2"
redis = "^3.4.1"

[tool.poetry.dev-dependencies]
pre-commit = "^2.2.0"
django-querycount = "^0.7.0"
factory_boy = "^2.12.0"
pytest = "^5.4.1"
pytest-django = "^3.9.0"
pytest-cov = "^2.8.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
